+++ spec/rubocop/cop/rspec/factory_bot/consistent_parentheses_style_spec.rb
@@ -2,8 +2,9 @@
 
 RSpec.describe RuboCop::Cop::RSpec::FactoryBot::ConsistentParenthesesStyle do
   let(:cop_config) do
-    { 'EnforcedStyle' => enforced_style }
+    { 'EnforcedStyle' => enforced_style, 'ExplicitOnly' => explicit_only }
   end
+  let(:explicit_only) { false }
 
   context 'when EnforcedStyle is :enforce_parentheses' do
     let(:enforced_style) { :require_parentheses }
@@ -380,6 +381,56 @@ RSpec.describe RuboCop::Cop::RSpec::FactoryBot::ConsistentParenthesesStyle do
       expect_no_offenses(<<~RUBY)
         generate()
         generate(:foo, :bar)
+      RUBY
+    end
+  end
+
+  context 'when ExplicitOnly is false' do
+    let(:enforced_style) { :require_parentheses }
+    let(:explicit_only) { false }
+
+    it 'register an offense when using `create` with an explicit receiver' do
+      expect_offense(<<~RUBY)
+        FactoryBot.create :user
+                   ^^^^^^ Prefer method call with parentheses
+      RUBY
+
+      expect_correction(<<~RUBY)
+        FactoryBot.create(:user)
+      RUBY
+    end
+
+    it 'register an offense when using `create` with no explicit receiver' do
+      expect_offense(<<~RUBY)
+        create :user
+        ^^^^^^ Prefer method call with parentheses
+      RUBY
+
+      expect_correction(<<~RUBY)
+        create(:user)
+      RUBY
+    end
+  end
+
+  context 'when ExplicitOnly is true' do
+    let(:enforced_style) { :require_parentheses }
+    let(:explicit_only) { true }
+
+    it 'register an offense when using `create` with an explicit receiver' do
+      expect_offense(<<~RUBY)
+        FactoryBot.create :user
+                   ^^^^^^ Prefer method call with parentheses
+      RUBY
+
+      expect_correction(<<~RUBY)
+        FactoryBot.create(:user)
+      RUBY
+    end
+
+    it 'dose not register an offense when using `create` ' \
+       'with no explicit receiver' do
+      expect_no_offenses(<<~RUBY)
+        create :user
       RUBY
     end
   end